requirements.txt
streamlit
PyGithub 
openai 
langchain 
langchain-openai
markdown